+++ Vic2ToHoI4/Source/HOI4World/HoI4Localisation.cpp
@@ -6,7 +6,7 @@
 #include "OSCompatibilityLayer.h"
 #include "States/HoI4State.h"
 #include "States/HoI4States.h"
-#include "V2World/Country.h"
+#include "V2World/Countries/Country.h"
 #include "V2World/Localisations/Vic2Localisations.h"
 #include "V2World/States/State.h"
 #include "V2World/States/StateDefinitions.h"
