+++ Vic2ToHoI4/Source/Mappers/GovernmentMapper.cpp
@@ -1,7 +1,7 @@
 #include "GovernmentMapper.h"
 #include "Log.h"
 #include "ParserHelpers.h"
-#include "V2World/Country.h"
+#include "V2World/Countries/Country.h"
