+++ Vic2ToHoI4/Source/OutHoi4/OutHoi4Country.cpp
@@ -19,7 +19,7 @@
 #include "OSCompatibilityLayer.h"
 #include "OutFocusTree.h"
 #include "OutTechnologies.h"
-#include "V2World/Country.h"
+#include "V2World/Countries/Country.h"
 #include <string>
