+++ Vic2ToHoI4/Source/V2World/Countries/Country.h
@@ -5,22 +5,22 @@
 
 #include "Ai/Vic2AI.h"
 #include "Color.h"
-#include "Culture/CultureGroups.h"
 #include "Date.h"
-#include "Diplomacy/Relations.h"
-#include "Diplomacy/RelationsFactory.h"
-#include "Localisations/Vic2Localisations.h"
-#include "Military/Army.h"
-#include "Military/Leaders/Leader.h"
-#include "Military/Leaders/LeaderFactory.h"
 #include "Parser.h"
-#include "Politics/Party.h"
-#include "Provinces/Province.h"
-#include "States/State.h"
-#include "States/StateDefinitions.h"
-#include "States/StateFactory.h"
-#include "Technology/Inventions.h"
-#include "Wars/War.h"
+#include "V2World/Culture/CultureGroups.h"
+#include "V2World/Diplomacy/Relations.h"
+#include "V2World/Diplomacy/RelationsFactory.h"
+#include "V2World/Localisations/Vic2Localisations.h"
+#include "V2World/Military/Army.h"
+#include "V2World/Military/Leaders/Leader.h"
+#include "V2World/Military/Leaders/LeaderFactory.h"
+#include "V2World/Politics/Party.h"
+#include "V2World/Provinces/Province.h"
+#include "V2World/States/State.h"
+#include "V2World/States/StateDefinitions.h"
+#include "V2World/States/StateFactory.h"
+#include "V2World/Technology/Inventions.h"
+#include "V2World/Wars/War.h"
 #include <map>
 #include <memory>
 #include <optional>
