+++ Vic2ToHoI4/Source/V2World/States/State.cpp
@@ -1,5 +1,5 @@
 #include "State.h"
-#include "V2World/Country.h"
+#include "V2World/Countries/Country.h"
 #include "V2World/Provinces/Province.h"
 #include <list>
