+++ Vic2ToHoI4/Source/V2World/World.cpp
@@ -1,7 +1,7 @@
 #include "World.h"
-#include "CommonCountryData.h"
 #include "Configuration.h"
-#include "Country.h"
+#include "Countries/Country.h"
+#include "Countries/CommonCountryData.h"
 #include "Ai/Vic2AI.h"
 #include "Culture/CultureGroupsFactory.h"
 #include "Date.h"
