+++ Vic2ToHoI4Tests/Vic2ToHoI4Tests-Google.vcxproj
@@ -285,10 +285,10 @@
     <ClCompile Include="..\Vic2ToHoI4\Source\OutHoi4\Sounds\OutSounds.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\OutHoi4\States\OutHoI4State.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\OutHoi4\States\OutHoI4States.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Countries\CommonCountryData.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Countries\Country.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Ai\AIStrategy.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Ai\Vic2AI.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\V2World\CommonCountryData.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Country.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Culture\CultureGroupFactory.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Culture\CultureGroups.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Culture\CultureGroupsFactory.cpp" />
